Split out a base driver from the cri & extension API adapters.
